+++ tilesets/generate_tiles.py
@@ -1,5 +1,5 @@
 import base64
-import tilesets.bigwig_tiles as bwt
+#import tilesets.bigwig_tiles as bwt
 import clodius.db_tiles as cdt
 import clodius.hdf_tiles as hdft
 import collections as col
@@ -7,6 +7,7 @@ import collections as col
 import clodius.tiles.bam as ctb
 import clodius.tiles.beddb as hgbe
 import clodius.tiles.bigwig as hgbi
+import clodius.tiles.bigbed as hgbb
 import clodius.tiles.cooler as hgco
 import clodius.tiles.geo as hggo
 import clodius.tiles.imtiles as hgim
@@ -504,6 +505,9 @@ def generate_tiles(tileset_tile_ids):
     elif tileset.filetype == 'bigwig':
         chromsizes = get_chromsizes(tileset)
         return hgbi.tiles(tileset.datafile.path, tile_ids, chromsizes=chromsizes)
+    elif tileset.filetype == 'bigbed':
+        chromsizes = get_chromsizes(tileset)
+        return hgbb.tiles(tileset.datafile.path, tile_ids, chromsizes=chromsizes)
     elif tileset.filetype == 'multivec':
         return generate_1d_tiles(
                 tileset.datafile.path,